reafactor: Split infra domain into repo
